Refactor launch files for Flexiv robot integration
